+++ tests/clickhouse_table_engine/test_base.py
@@ -6,6 +6,7 @@ from importlib import import_module
 
 from django.apps import apps
 from django.db import connection, connections, migrations, models
+from django.db import models as django_models
 from django.db.migrations.migration import Migration
 from django.db.migrations.recorder import MigrationRecorder
 from django.db.migrations.state import ProjectState
@@ -13,7 +14,7 @@ from django.test import TransactionTestCase
 from django.test.utils import extend_sys_path
 from django.utils.module_loading import module_dir
 
-from clickhouse_backend.models.indexes import Index, Set
+from clickhouse_backend import models
 
 
 class MigrationTestBase(TransactionTestCase):
@@ -22,7 +23,7 @@ class MigrationTestBase(TransactionTestCase):
     """
 
     available_apps = ["migrations"]
-    databases = {"default", "other"}
+    databases = {"default", "s1r2", "s2r1", "other"}
 
     def tearDown(self):
         # Reset applied-migrations state.
@@ -247,166 +248,153 @@ class OperationTestBase(MigrationTestBase):
         Makes a test state using set_up_test_model and returns the
         original state and the state after the migration is applied.
         """
-        project_state = self.set_up_test_model(app_label, **kwargs)
+        project_state = self.set_up_distributed_model(app_label, **kwargs)
         new_state = project_state.clone()
         operation.state_forwards(app_label, new_state)
         return project_state, new_state
 
-    def set_up_test_model(
+    def set_up_distributed_model(
         self,
         app_label,
-        second_model=False,
-        third_model=False,
         index=False,
-        multicol_index=False,
         related_model=False,
-        mti_model=False,
-        proxy_model=False,
-        manager_model=False,
-        options=False,
-        db_table=None,
-        index_together=False,  # RemovedInDjango51Warning.
+        multicol_index=False,
+        index_together=False,
         constraints=None,
         indexes=None,
     ):
         """Creates a test model state and database table."""
-        # Make the "current" state.
-        model_options = {
-            "swappable": "TEST_SWAP_MODEL",
-            # RemovedInDjango51Warning.
-            "indexes": [
-                Index(
+        meta_indexes = (
+            [
+                models.Index(
                     fields=("weight", "pink"),
                     name="weight_pink_idx",
-                    type=Set(100),
+                    type=models.Set(100),
                     granularity=10,
                 )
             ]
             if index_together
-            else [],
-        }
-        if options:
-            model_options["permissions"] = [("can_groom", "Can groom")]
-        if db_table:
-            model_options["db_table"] = db_table
+            else []
+        )
+
         operations = [
             migrations.CreateModel(
                 "Pony",
                 [
-                    ("id", models.BigAutoField(primary_key=True)),
-                    ("pink", models.IntegerField(default=3)),
-                    ("weight", models.FloatField()),
+                    ("id", django_models.BigAutoField(primary_key=True)),
+                    ("pink", models.Int32Field(default=3)),
+                    ("weight", models.Float64Field()),
                 ],
-                options=model_options,
-            )
+                options={
+                    "indexes": meta_indexes,
+                    "engine": models.ReplicatedMergeTree(order_by='id'),
+                    "cluster": "cluster",
+                },
+            ),
+            migrations.CreateModel(
+                "PonyDistributed",
+                [
+                    ("id", django_models.BigAutoField(primary_key=True)),
+                    ("pink", models.Int32Field(default=3)),
+                    ("weight", models.Float64Field()),
+                ],
+                options={
+                    "indexes": meta_indexes,
+                    "engine": models.Distributed(
+                        "cluster", models.currentDatabase(), f"{app_label}_pony"
+                    ),
+                    "cluster": "cluster",
+                },
+            ),
         ]
-        if index:
-            operations.append(
-                migrations.AddIndex(
-                    "Pony",
-                    Index(
-                        fields=["pink"],
-                        name="pony_pink_idx",
-                        type=Set(100),
-                        granularity=10,
-                    ),
+
+        for name in ["pony", "ponydistributed"]:
+            if index:
+                operations.append(
+                    migrations.AddIndex(
+                        name,
+                        models.Index(
+                            fields=["pink"],
+                            name=f"{name}_pink_idx",
+                            type=models.Set(100),
+                            granularity=10,
+                        ),
+                    )
                 )
-            )
-        if multicol_index:
-            operations.append(
-                migrations.AddIndex(
-                    "Pony",
-                    Index(
-                        fields=["pink", "weight"],
-                        name="pony_test_idx",
-                        type=Set(100),
-                        granularity=10,
-                    ),
+            if multicol_index:
+                operations.append(
+                    migrations.AddIndex(
+                        name,
+                        models.Index(
+                            fields=["pink", "weight"],
+                            name=f"{name}_test_idx",
+                            type=models.Set(100),
+                            granularity=10,
+                        ),
+                    )
                 )
-            )
-        if indexes:
-            for index in indexes:
-                operations.append(migrations.AddIndex("Pony", index))
-        if constraints:
-            for constraint in constraints:
-                operations.append(migrations.AddConstraint("Pony", constraint))
-        if second_model:
-            operations.append(
-                migrations.CreateModel(
-                    "Stable",
-                    [
-                        ("id", models.BigAutoField(primary_key=True)),
-                    ],
-                )
-            )
-        if third_model:
-            operations.append(
-                migrations.CreateModel(
-                    "Van",
-                    [
-                        ("id", models.BigAutoField(primary_key=True)),
-                    ],
-                )
-            )
+            if indexes:
+                for index in indexes:
+                    operations.append(migrations.AddIndex(name, index))
+            if constraints:
+                for constraint in constraints:
+                    operations.append(migrations.AddConstraint(name, constraint))
         if related_model:
             operations.append(
                 migrations.CreateModel(
                     "Rider",
                     [
-                        ("id", models.BigAutoField(primary_key=True)),
-                        ("pony", models.ForeignKey("Pony", models.CASCADE)),
+                        ("id", django_models.BigAutoField(primary_key=True)),
+                        (
+                            "pony",
+                            django_models.ForeignKey("Pony", django_models.CASCADE),
+                        ),
                         (
                             "friend",
-                            models.ForeignKey("self", models.CASCADE, null=True),
-                        ),
-                    ],
-                )
-            )
-        if mti_model:
-            operations.append(
-                migrations.CreateModel(
-                    "ShetlandPony",
-                    fields=[
-                        (
-                            "pony_ptr",
-                            models.OneToOneField(
-                                "Pony",
-                                models.CASCADE,
-                                auto_created=True,
-                                parent_link=True,
-                                primary_key=True,
-                                to_field="id",
-                                serialize=False,
+                            django_models.ForeignKey(
+                                "self", django_models.CASCADE, null=True
                             ),
                         ),
-                        ("cuteness", models.IntegerField(default=1)),
                     ],
-                    bases=["%s.Pony" % app_label],
+                    options={
+                        "engine": models.ReplicatedMergeTree(order_by='id'),
+                        "cluster": "cluster",
+                    },
                 )
             )
-        if proxy_model:
             operations.append(
                 migrations.CreateModel(
-                    "ProxyPony",
-                    fields=[],
-                    options={"proxy": True},
-                    bases=["%s.Pony" % app_label],
+                    "RiderDistributed",
+                    [
+                        ("id", django_models.BigAutoField(primary_key=True)),
+                        (
+                            "pony",
+                            django_models.ForeignKey(
+                                "PonyDistributed", django_models.CASCADE
+                            ),
+                        ),
+                        (
+                            "friend",
+                            django_models.ForeignKey(
+                                "self", django_models.CASCADE, null=True
+                            ),
+                        ),
+                    ],
+                    options={
+                        "engine": models.Distributed(
+                            "cluster", models.currentDatabase(), f"{app_label}_rider"
+                        ),
+                        "cluster": "cluster",
+                    },
                 )
             )
-        if manager_model:
-            from .models import FoodManager, FoodQuerySet
 
-            operations.append(
-                migrations.CreateModel(
-                    "Food",
-                    fields=[
-                        ("id", models.BigAutoField(primary_key=True)),
-                    ],
-                    managers=[
-                        ("food_qs", FoodQuerySet.as_manager()),
-                        ("food_mgr", FoodManager("a", "b")),
-                        ("food_mgr_kwargs", FoodManager("x", "y", 3, 4)),
-                    ],
-                )
-            )
         return self.apply_operations(app_label, ProjectState(), operations)
+
+    def assertTableNotExistsCluster(self, table, dbs=("default", "s1r2", "s2r1")):
+        for db in dbs:
+            self.assertTableNotExists(table, db)
+
+    def assertTableExistsCluster(self, table, dbs=("default", "s1r2", "s2r1")):
+        for db in dbs:
+            self.assertTableExists(table, db)
